requirements.txt
google-generativeai
protobuf~=5.28.2
gunicorn==22.0.0
flask
jsonify
re
os
werkzeug.utils